+++ src/Node.java
@@ -1,13 +1,13 @@
 public class Node {
-	Double valor;
+	int valor;
 	Node esquerda;
 	Node direita;
 
-	public Node(Double valor) {
+	public Node(int valor) {
 		this.valor = valor;
 	}
 
-	public Node(Double valor, Node esquerda, Node direita) {
+	public Node(int valor, Node esquerda, Node direita) {
 		this.valor = valor;
 		this.esquerda = esquerda;
 		this.direita = direita;
